[댓글 기능] 션 미션 제출합니다. (#141)
* Add : 객체관계 매핑 수업내용 추가

* Add : 게시글 기능에 사용자 인증 추가

* Add : 댓글 작성 및 조회 기능 추가

* Add : 댓글 삭제 기능 추가

* Add : 게시글 삭제 기능에 사용자 인증 추가

* Refactor : 댓글창을 textarea -> tui editor로 수정

* Add : 댓글 수정 기능 추가

* Refactor : CommentController의 RequestMapping 중복제거

* Add : 댓글 수정, 삭제에 사용자 인증 추가

* Add : 댓글 작성 테스트 추가

* Mod : Model 패키지 분리

* Add : Article, Comment에 생성/수정시간 필드 추가

- with Spring JPA Audit

* Mod : Article의 contents를 @Lob 설정

* Refactor : Article 수정/삭제의 사용자 인증관련 리팩토링

* Refactor : Comment 수정/삭제의 사용자 인증관련 리팩토링

* Refactor : UserController의 url convention 수정

* Modify : 상수처리

* Refactor : User 수정/삭제의 사용자 인증관련 리팩토링, Exception 이름 수정

* Refactor : 로그인 관련 로직 리팩토링

* Refactor : User 수정/삭제 리팩토링에 따른 UserAuthInterceptor 제거

* Refactor : custom Exception 처리

* Refactor : Dto 네이밍방식 수정

+++ src/main/java/techcourse/myblog/domain/service/ArticleService.java
@@ -3,8 +3,10 @@ package techcourse.myblog.domain.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.exception.ArticleNotFoundException;
+import techcourse.myblog.domain.exception.MisMatchUserException;
+import techcourse.myblog.domain.model.Article;
+import techcourse.myblog.domain.model.User;
 import techcourse.myblog.domain.repository.ArticleRepository;
 
 import java.util.List;
@@ -25,25 +27,40 @@ public class ArticleService {
         return unmodifiableList(articleRepository.findAll());
     }
 
+    @Transactional(readOnly = true)
+    public Article findById(long id) {
+        return articleRepository.findById(id)
+                .orElseThrow(() -> new ArticleNotFoundException("해당하는 게시글을 찾지 못했습니다."));
+    }
+
+    @Transactional(readOnly = true)
+    public Article findByIdAsAuthor(long id, User user) {
+        Article article = findById(id);
+        if (!article.isAuthor(user)) {
+            throw new MisMatchUserException("작성자만 접근가능합니다.");
+        }
+        return article;
+    }
+
     @Transactional
     public Article save(Article article) {
         return articleRepository.save(article);
     }
 
     @Transactional
-    public Article findById(long id) {
-        return articleRepository.findById(id).orElseThrow(() -> new ArticleNotFoundException("해당하는 게시글을 찾지 못했습니다."));
+    public Article updateByIdAsAuthor(long id, Article articleToUpdate) {
+        Article article = findById(id);
+        if (!article.isAuthor(articleToUpdate.getAuthor())) {
+            throw new MisMatchUserException("작성자만 접근가능합니다.");
+        }
+        return article.update(articleToUpdate);
     }
 
     @Transactional
-    public Article update(long id, Article articleToUpdate) {
-        Article originArticle = articleRepository.findArticleById(id);
-        originArticle.update(articleToUpdate);
-        return originArticle;
-    }
-
-    @Transactional
-    public void deleteById(long id) {
+    public void deleteByIdAsAuthor(long id, User user) {
+        if (!findById(id).isAuthor(user)) {
+            throw new MisMatchUserException("작성자만 접근가능합니다.");
+        }
         articleRepository.deleteById(id);
     }
 }
